Change: nRF Provisioner "App" now defines the Intro sheet
This is no longer in ScannerView, since this sheet is meant to be in App Launch, not when specifically attempting to provision via BLE as it was before. To do this, we defined a new "ViewModel" for the App, and removed the properties associated to the sheet from the old Scanner View Model, to put them in the new one.