Work-In-Progress: mDNS / Bonjour Service Discovery
Rather than sending requests directly to an IP Address, we'd like to 'discover' the Service in its own network. We can't do much past this point because current firmware does not advertise any Bonjour Service, but what we can say is that this seems to be doing the browsing part, at least.