Changed Titles for "Forget" and "Set / Update" Configuration
"Configuration" is no longer in the button text, so the buttons don't need to be as big and the UI looks better in smaller devices. Which is a rare thing given how the smallest iPhone is now quite big, but might be a thing for some users, so it's worth still keeping this in check.