Updated APWiFiScanView Layout
The icon is now aligned with the "Device Status" Section as well as the "Access Point" Section below it, they're all left-aligned. Also, the Channel & Security info on each scan result are now colored using the .secondaryColor (my God, what a save that we worked on that late-day yesterday). What's left now is the right-alignment of the checkmark.