RSSIView and Selected AP Result Checkmark are now "center aligned" within AP Title Baseline
I don't know how much better to write that. This is not perfect, because the RSSI is our custom drawing, and it's not size consistent in that there's "empty space" when the bar is not filled, so it doesn't look consistent with our padding design. So we either fix the bars, or we switch to SwiftUI SF Symbols for the bar drawings.